adjust speed

+++ Client/gui/paint/GamePainter.java
@@ -32,7 +32,7 @@ public class GamePainter {
         g2d.setColor(Color.BLACK);
         g2d.setRenderingHint(RenderingHints.KEY_TEXT_ANTIALIASING, RenderingHints.VALUE_TEXT_ANTIALIAS_ON);
         g2d.setFont(new Font("Impact", Font.BOLD, 30));
-        int level = (int)(lineCnt / 5.0) + 1;
+        int level = lineCnt / 5 + 1;
         if (level >= 20) level = 20;
         g2d.drawString("Lv:   " + level, 92 + pos, 66);
         g2d.drawString("SCORE:   " + String.format("%06d", score), 198 + pos, 66);
